Updated mod to 1.7
Completely rewrote mod from scratch. Should be a lot better in game and
code wise than the old one. Also has mod support for other mods. Plus it
adds a ton of new features; including throwable ingots and bricks that
can be customised though NBT. This means any in game item can now be
throwable, as well as have the damage set, if it explodes and how big it
explodes, the size of the brick, etc.